server_app/requirements.txt
aiofiles==22.1.0
anyio==3.6.2
attrs==22.2.0
click==8.1.3
exceptiongroup==1.1.0
fastapi==0.89.1
fastapi-sessions==0.3.2
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
psycopg2==2.9.5
pydantic==1.10.4
pydicom==2.3.1
pytest==7.2.1
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
